Moved to RecommendationsController
